feat(blog): remove unused image tags from acorn post
Removed multiple empty img tags that were not being used in the post content.
These tags did not have any src or alt attributes populated and appeared to be
placeholders that were never filled in. The removal cleans up the HTML structure
and improves the overall readability of the markdown file.